getEstimates

This function calculates the point estimates and credible intervals per cohort, as well as estimates of the biases and the mean squared errors of the point estimates per cohort.

getEstimates(
  analyses_list,
  add_parameters = NULL,
  point_estimator = "median",
  alpha_level = 0.05
)

Arguments

analyses_list

An object of class analysis_list, as created with performAnalyses

add_parameters

A vector of strings naming additional parameters from the Bayesian hierarchical models, e.g. c('mu', 'tau'). If NULL, no additional parameters will be evaluated, Default: NULL

point_estimator

A string indicating the type of estimator used for calculation of bias and MSE. Must be one of 'median' or 'mean'

alpha_level

A numeric in (0, 1) for the level of the credible interval. Only values corresponding to quantiles saved in performAnalyses will work, Default: 0.05

Value

A named list of matrices of estimates of cohort-level parameters and credible intervals. Estimates of bias and MSE are included for cohort-level posterior estimates of simulated trials.

Details

For binary endpoints, bias and MSE will only be calculated for posterior response-rate estimates of simulated trials. For continuous endpoints, bias and MSE will only be calculated for posterior cohort-mean estimates of simulated trials. For additional parameters, bias and MSE will not be calculated.

Possible additional parameters for the Bayesian hierarchical models are c('mu', 'tau') for 'berry', 'exnex', 'exnex_mix', 'exnex_adj', and 'exnex_adj_mix'. The ExNex-based models can also access posterior weights. paste0("w_", seq_len(n_cohorts)).

See also

Author

Stephan Wojciekowski

Examples

## Binary endpoint example
scenarios_list <- simulateScenarios(
  n_subjects_list     = list(c(10, 20, 30)),
  response_rates_list = list(c(0.1, 0.2, 3)),
  n_trials            = 10)

analyses_list <- performAnalyses(
  scenario_list       = scenarios_list,
  target_rates        = c(0.1, 0.1, 0.1),
  calc_differences    = matrix(c(3, 2, 2, 1), ncol = 2),
  n_mcmc_iterations   = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 12-Jul-2026 Performing Analyses
#>          Analyzing 1 scenario (9 unique trial realizations)
#>    15:54 - with Berry ...........
#>              finished after 0.1 secs.
#>    15:54 - with Exnex ...........
#>              finished after 0.1 secs.
#>    15:54 - with Exnex_adj .......
#>              finished after 0.1 secs.
#>    15:54 - with Exnex_adj_mix ...
#>              finished after 0.1 secs.
#>    15:54 - with Exnex_mix .......
#>              finished after 0.1 secs.
#>    15:54 - with Pooled ..........
#>              finished after 0.1 secs.
#>    15:54 - with Stratified ......
#>              finished after 0.1 secs.
#>    15:54 - with Stratified_mix ...
#>              finished after 0.1 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

getEstimates(analyses_list)
#> $berry
#> $berry$scenario_1
#>                  Mean         SD        2.5%         50%      97.5%
#> p_1        0.09327857 0.05890487  0.01592167  0.08220715 0.23171964
#> p_2        0.16917770 0.06671241  0.06645957  0.16143858 0.31528334
#> p_3        0.10871577 0.04988839  0.03591730  0.10019479 0.21766061
#> p_diff_32 -0.06046193 0.07373014 -0.21056838 -0.05420452 0.07324458
#> p_diff_21  0.07589913 0.08003576 -0.07431704  0.07259940 0.22994007
#>                    Bias          MSE
#> p_1       -0.0177928480 0.0008846444
#> p_2       -0.0385614190 0.0063941857
#> p_3        0.0001947913 0.0001952927
#> p_diff_32 -0.1542045190 0.0276971925
#> p_diff_21  0.1725993982 0.0356923441
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>                  Mean         SD         2.5%         50%      97.5%
#> p_1        0.06189621 0.05992801  0.006116128  0.04165491 0.21174070
#> p_2        0.19128835 0.07558310  0.075166720  0.18706150 0.35741189
#> p_3        0.10352106 0.04960930  0.032429030  0.09554788 0.20963980
#> p_diff_32 -0.08776729 0.08791196 -0.268828753 -0.08382263 0.06814992
#> p_diff_21  0.12939214 0.09779941 -0.057351657  0.13291493 0.31190742
#>                   Bias          MSE
#> p_1       -0.058345094 5.441800e-03
#> p_2       -0.012938499 8.520826e-03
#> p_3       -0.004452121 8.698073e-05
#> p_diff_32 -0.183822627 4.068814e-02
#> p_diff_21  0.232914927 6.657872e-02
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>                  Mean         SD         2.5%         50%      97.5%
#> p_1        0.06599258 0.06124893  0.004954785  0.04495083 0.22641384
#> p_2        0.18750115 0.08116955  0.068714055  0.17503944 0.37374195
#> p_3        0.10278895 0.05209479  0.027249792  0.09560881 0.21538310
#> p_diff_32 -0.08471221 0.09939046 -0.291731610 -0.07874853 0.08713122
#> p_diff_21  0.12150857 0.10588555 -0.079520986  0.11903320 0.32797248
#>                   Bias          MSE
#> p_1       -0.055049166 0.0054704976
#> p_2       -0.024960559 0.0077270249
#> p_3       -0.004391192 0.0001018482
#> p_diff_32 -0.178748528 0.0382317058
#> p_diff_21  0.219033195 0.0592413665
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>                  Mean         SD         2.5%         50%      97.5%
#> p_1        0.06137817 0.06144668  0.006185392  0.03916864 0.21993037
#> p_2        0.17643786 0.07632208  0.062508189  0.16671907 0.35265947
#> p_3        0.10386791 0.05125757  0.027279997  0.09789538 0.21596214
#> p_diff_32 -0.07256995 0.08908818 -0.275019812 -0.06262672 0.08156056
#> p_diff_21  0.11505969 0.09582819 -0.064750368  0.11660333 0.30194688
#>                   Bias          MSE
#> p_1       -0.060831357 0.0059627069
#> p_2       -0.033280928 0.0076306838
#> p_3       -0.002104617 0.0001427775
#> p_diff_32 -0.162626723 0.0317096253
#> p_diff_21  0.216603325 0.0574265552
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>                  Mean         SD        2.5%         50%      97.5%
#> p_1        0.06548668 0.06417073  0.00590155  0.04476692 0.22959784
#> p_2        0.18526289 0.07530676  0.06933054  0.17821220 0.35488319
#> p_3        0.10606990 0.05184953  0.02609442  0.09876290 0.21961239
#> p_diff_32 -0.07919300 0.08994135 -0.26883084 -0.06943676 0.07091265
#> p_diff_21  0.11977621 0.10259688 -0.08325212  0.12376227 0.31252548
#>                   Bias          MSE
#> p_1       -0.055233075 0.0051739318
#> p_2       -0.021787805 0.0097536330
#> p_3       -0.001237103 0.0001308545
#> p_diff_32 -0.169436762 0.0357568063
#> p_diff_21  0.223762270 0.0635578110
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>                Mean         SD       2.5%       50%     97.5%        Bias
#> p_1       0.1229508 0.04137543 0.05438267 0.1188289 0.2147772  0.01882888
#> p_2       0.1229508 0.04137543 0.05438267 0.1188289 0.2147772 -0.08117112
#> p_3       0.1229508 0.04137543 0.05438267 0.1188289 0.2147772  0.01882888
#> p_diff_32 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000 -0.10000000
#> p_diff_21 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000  0.10000000
#>                   MSE
#> p_1       0.001159639
#> p_2       0.007393863
#> p_3       0.001159639
#> p_diff_32 0.010000000
#> p_diff_21 0.010000000
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>                  Mean         SD         2.5%         50%     97.5%
#> p_1        0.04545455 0.04784504  0.003607962  0.03249941 0.1730025
#> p_2        0.19523810 0.08074703  0.067322378  0.18557978 0.3768327
#> p_3        0.10000000 0.05303301  0.022626655  0.09142896 0.2253392
#> p_diff_32 -0.10460532 0.09867043 -0.307144339 -0.09817760 0.0652012
#> p_diff_21  0.15416793 0.09887542 -0.029284277  0.14739901 0.3423886
#>                   Bias          MSE
#> p_1       -0.067500587 7.960856e-03
#> p_2       -0.014420217 9.292575e-03
#> p_3       -0.008571039 7.346271e-05
#> p_diff_32 -0.198177598 4.765760e-02
#> p_diff_21  0.247399012 7.888757e-02
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>                  Mean         SD         2.5%         50%      97.5%
#> p_1        0.05220035 0.05375643  0.005452479  0.03702717 0.20722666
#> p_2        0.20580708 0.08310697  0.078580497  0.20064213 0.38504508
#> p_3        0.10594143 0.05518165  0.026281044  0.09527343 0.22103543
#> p_diff_32 -0.10128676 0.09972699 -0.316205551 -0.09344287 0.07910224
#> p_diff_21  0.14804714 0.10128996 -0.038397217  0.15100799 0.32664838
#>                    Bias          MSE
#> p_1       -0.0629728310 0.0080182818
#> p_2        0.0006421326 0.0101112529
#> p_3       -0.0047265698 0.0001014202
#> p_diff_32 -0.1934428718 0.0459036792
#> p_diff_21  0.2510079876 0.0816193375
#> 
#> 
getEstimates(analyses_list   = analyses_list,
             add_parameters  = c("mu", "tau", "w_1", "w_2", "w_3"),
             point_estimator = "mean",
             alpha_level     = 0.1)
#> $berry
#> $berry$scenario_1
#>                   Mean         SD          5%         50%        95%
#> p_1        0.093278571 0.05890487  0.02009166  0.08220715 0.19787008
#> p_2        0.169177705 0.06671241  0.07694027  0.16143858 0.28745962
#> p_3        0.108715773 0.04988839  0.04242673  0.10019479 0.19487571
#> p_diff_32 -0.060461931 0.07373014 -0.18950765 -0.05420452 0.05334238
#> p_diff_21  0.075899133 0.08003576 -0.03786932  0.07259940 0.20495168
#> mu         0.009066839 0.70602163 -1.10324651  0.05263879 0.97388302
#> tau        0.752549176 0.50032828  0.16725446  0.64790895 1.72185282
#>                   Bias          MSE
#> p_1       -0.006721429 0.0006422741
#> p_2       -0.030822295 0.0057559787
#> p_3        0.008715773 0.0002694794
#> p_diff_32 -0.160461931 0.0297106596
#> p_diff_21  0.175899133 0.0366373493
#> mu                  NA           NA
#> tau                 NA           NA
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>                  Mean         SD           5%         50%        95%
#> p_1        0.06189621 0.05992801  0.008569334  0.04165491 0.17510404
#> p_2        0.19128835 0.07558310  0.083753318  0.18706150 0.32507748
#> p_3        0.10352106 0.04960930  0.038456790  0.09554788 0.19217096
#> p_diff_32 -0.08776729 0.08791196 -0.233315622 -0.08382263 0.04510189
#> p_diff_21  0.12939214 0.09779941 -0.025770545  0.13291493 0.27956916
#> mu        -2.38747582 1.90116055 -5.949096570 -2.24127018 0.68432246
#> tau        0.71788277 0.51688250  0.103363093  0.62306836 1.72611130
#> w_1        0.47910448 0.48561688  0.000000000  0.50000000 1.00000000
#> w_2        0.51791045 0.47123543  0.000000000  0.60000000 1.00000000
#> w_3        0.54925373 0.48456872  0.000000000  0.50000000 1.00000000
#>                   Bias          MSE
#> p_1       -0.038103793 3.369414e-03
#> p_2       -0.008711651 8.151138e-03
#> p_3        0.003521062 8.249085e-05
#> p_diff_32 -0.187767287 4.266299e-02
#> p_diff_21  0.229392141 6.618215e-02
#> mu                  NA           NA
#> tau                 NA           NA
#> w_1                 NA           NA
#> w_2                 NA           NA
#> w_3                 NA           NA
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>                  Mean         SD           5%         50%        95%
#> p_1        0.06599258 0.06124893  0.007022074  0.04495083 0.18872110
#> p_2        0.18750115 0.08116955  0.080590673  0.17503944 0.33551610
#> p_3        0.10278895 0.05209479  0.034421109  0.09560881 0.19447923
#> p_diff_32 -0.08471221 0.09939046 -0.251645622 -0.07874853 0.06167748
#> p_diff_21  0.12150857 0.10588555 -0.047437086  0.11903320 0.28804435
#> mu        -0.06134294 1.65184864 -2.786785902  0.12456700 2.30881689
#> tau        0.76534447 0.54291575  0.132623034  0.64808025 1.81240302
#> w_1        0.44925373 0.49066200  0.000000000  0.35000000 1.00000000
#> w_2        0.54925373 0.48652810  0.000000000  0.70000000 1.00000000
#> w_3        0.57462687 0.48647571  0.000000000  0.70000000 1.00000000
#>                   Bias          MSE
#> p_1       -0.034007420 3.700993e-03
#> p_2       -0.012498847 6.906162e-03
#> p_3        0.002788946 3.531545e-05
#> p_diff_32 -0.184712206 4.021777e-02
#> p_diff_21  0.221508572 6.169639e-02
#> mu                  NA           NA
#> tau                 NA           NA
#> w_1                 NA           NA
#> w_2                 NA           NA
#> w_3                 NA           NA
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>                  Mean         SD           5%         50%        95%
#> p_1        0.06137817 0.06144668  0.007888497  0.03916864 0.17927423
#> p_2        0.17643786 0.07632208  0.073234844  0.16671907 0.30708330
#> p_3        0.10386791 0.05125757  0.034634973  0.09789538 0.19517633
#> p_diff_32 -0.07256995 0.08908818 -0.216433938 -0.06262672 0.05923204
#> p_diff_21  0.11505969 0.09582819 -0.031754035  0.11660333 0.26105875
#> mu         0.11458265 1.10083839 -1.681830657  0.14450972 1.87243177
#> tau        0.72426123 0.51213967  0.118154304  0.60894695 1.70989166
#> w_1        0.40820896 0.45929168  0.000000000  0.30000000 1.00000000
#> w_2        0.68880597 0.44954999  0.000000000  1.00000000 1.00000000
#> w_3        0.70000000 0.44466477  0.000000000  0.90000000 1.00000000
#>                   Bias          MSE
#> p_1       -0.038621827 0.0036830115
#> p_2       -0.023562139 0.0071924250
#> p_3        0.003867912 0.0001028057
#> p_diff_32 -0.172569950 0.0357865411
#> p_diff_21  0.215059688 0.0578534051
#> mu                  NA           NA
#> tau                 NA           NA
#> w_1                 NA           NA
#> w_2                 NA           NA
#> w_3                 NA           NA
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>                  Mean         SD           5%         50%         95%
#> p_1        0.06548668 0.06417073  0.008244691  0.04476692  0.19312264
#> p_2        0.18526289 0.07530676  0.083911392  0.17821220  0.31552523
#> p_3        0.10606990 0.05184953  0.034070503  0.09876290  0.20024302
#> p_diff_32 -0.07919300 0.08994135 -0.240598202 -0.06943676  0.04761936
#> p_diff_21  0.11977621 0.10259688 -0.047950557  0.12376227  0.27495068
#> mu        -2.61656699 1.57436273 -5.845485622 -2.33184297 -0.80860521
#> tau        0.58473269 0.49865174  0.084387133  0.42308637  1.52003577
#> w_1        0.50074627 0.48076693  0.000000000  0.40000000  1.00000000
#> w_2        0.47313433 0.46508608  0.000000000  0.60000000  1.00000000
#> w_3        0.62164179 0.47716032  0.000000000  0.85000000  1.00000000
#>                   Bias          MSE
#> p_1       -0.034513315 0.0033843263
#> p_2       -0.014737106 0.0085735838
#> p_3        0.006069899 0.0001465864
#> p_diff_32 -0.179192996 0.0398052264
#> p_diff_21  0.219776210 0.0626813524
#> mu                  NA           NA
#> tau                 NA           NA
#> w_1                 NA           NA
#> w_2                 NA           NA
#> w_3                 NA           NA
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>                Mean         SD         5%       50%       95%        Bias
#> p_1       0.1229508 0.04137543 0.06259345 0.1188289 0.1974055  0.02295082
#> p_2       0.1229508 0.04137543 0.06259345 0.1188289 0.1974055 -0.07704918
#> p_3       0.1229508 0.04137543 0.06259345 0.1188289 0.1974055  0.02295082
#> p_diff_32 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000 -0.10000000
#> p_diff_21 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000  0.10000000
#>                   MSE
#> p_1       0.001315059
#> p_2       0.006724895
#> p_3       0.001315059
#> p_diff_32 0.010000000
#> p_diff_21 0.010000000
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>                  Mean         SD           5%         50%        95%
#> p_1        0.04545455 0.04784504  0.005530457  0.03249941 0.13323852
#> p_2        0.19523810 0.08074703  0.081023892  0.18557978 0.34259328
#> p_3        0.10000000 0.05303301  0.029598756  0.09142896 0.19978101
#> p_diff_32 -0.10460532 0.09867043 -0.280967438 -0.09817760 0.04141252
#> p_diff_21  0.15416793 0.09887542  0.001910477  0.14739901 0.31624478
#>                   Bias         MSE
#> p_1       -0.054545455 0.007015611
#> p_2       -0.004761905 0.008589065
#> p_3        0.000000000 0.000000000
#> p_diff_32 -0.204605324 0.050230164
#> p_diff_21  0.254167934 0.084333197
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>                  Mean         SD           5%         50%        95%
#> p_1        0.05220035 0.05375643  0.007777511  0.03702717 0.17026068
#> p_2        0.20580708 0.08310697  0.092666079  0.20064213 0.35092713
#> p_3        0.10594143 0.05518165  0.031983152  0.09527343 0.20039096
#> p_diff_32 -0.10128676 0.09972699 -0.262610146 -0.09344287 0.04853145
#> p_diff_21  0.14804714 0.10128996 -0.004119931  0.15100799 0.30589427
#>                   Bias          MSE
#> p_1       -0.047799649 7.307385e-03
#> p_2        0.005807080 8.983913e-03
#> p_3        0.005941434 3.530063e-05
#> p_diff_32 -0.201286755 4.924372e-02
#> p_diff_21  0.248047140 8.191920e-02
#> 
#> 

outcome <- createTrial(
  n_subjects   = c(10, 20, 30),
  n_responders = c(1, 2, 3))

outcome_analysis <- performAnalyses(
  scenario_list       = outcome,
  target_rates        = c(0.1, 0.1, 0.1),
  n_mcmc_iterations   = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 12-Jul-2026 Performing Analyses
#>          Analyzing 1 scenario (1 unique trial realization)
#>    15:54 - with Berry ...........
#>              finished after 0.1 secs.
#>    15:54 - with Exnex ...........
#>              finished after 0.1 secs.
#>    15:54 - with Exnex_adj .......
#>              finished after 0.1 secs.
#>    15:54 - with Exnex_adj_mix ...
#>              finished after 0.1 secs.
#>    15:54 - with Exnex_mix .......
#>              finished after 0.1 secs.
#>    15:54 - with Pooled ..........
#>              finished after 0 secs.
#>    15:54 - with Stratified ......
#>              finished after 0 secs.
#>    15:54 - with Stratified_mix ...
#>              finished after 0 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

getEstimates(outcome_analysis)
#> $berry
#> $berry$scenario_1
#>          Mean         SD       2.5%       50%     97.5%
#> p_1 0.1175256 0.06087979 0.03494172 0.1084307 0.2337896
#> p_2 0.1143987 0.04624819 0.04029308 0.1160492 0.2083065
#> p_3 0.1134232 0.04659492 0.03790970 0.1058081 0.2012647
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>           Mean         SD       2.5%        50%     97.5%
#> p_1 0.09773350 0.06956449 0.01653333 0.07987617 0.2945066
#> p_2 0.09198751 0.04852005 0.02362535 0.08527208 0.2089434
#> p_3 0.09373415 0.04678252 0.02969117 0.08714735 0.2231015
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>          Mean         SD        2.5%        50%     97.5%
#> p_1 0.1014439 0.06549089 0.009655962 0.09951111 0.2537204
#> p_2 0.1030144 0.06183202 0.014677379 0.09333534 0.2282552
#> p_3 0.1042839 0.04825738 0.032063702 0.10274114 0.2089005
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>           Mean         SD        2.5%        50%     97.5%
#> p_1 0.10037715 0.07238333 0.009448109 0.09142011 0.2942932
#> p_2 0.08905221 0.04935013 0.010584831 0.08252995 0.1907366
#> p_3 0.09329984 0.04535222 0.025731479 0.08336735 0.2087757
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>           Mean         SD       2.5%        50%     97.5%
#> p_1 0.09603272 0.05981212 0.01179265 0.09055277 0.2386408
#> p_2 0.09944328 0.05425500 0.02791187 0.09676626 0.2276628
#> p_3 0.09925029 0.04503643 0.03714694 0.09252128 0.2067330
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>     Mean         SD       2.5%        50%     97.5%
#> p_1  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> p_2  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> p_3  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>     Mean         SD        2.5%        50%     97.5%
#> p_1  0.1 0.08660254 0.003720276 0.07636130 0.3240643
#> p_2  0.1 0.06396021 0.014000117 0.08739591 0.2559472
#> p_3  0.1 0.05303301 0.022626655 0.09142896 0.2253392
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>          Mean         SD       2.5%        50%     97.5%
#> p_1 0.1155437 0.09506287 0.00448041 0.09553074 0.2767190
#> p_2 0.1085965 0.06762380 0.01276664 0.09273470 0.2379104
#> p_3 0.1059414 0.05518165 0.02063231 0.09774616 0.2469729
#> 
#> 
getEstimates(analyses_list  = outcome_analysis,
             add_parameters = c("mu", "w_1", "w_2", "w_3"))
#> $berry
#> $berry$scenario_1
#>           Mean         SD        2.5%       50%     97.5%
#> p_1 0.11752561 0.06087979  0.03494172 0.1084307 0.2337896
#> p_2 0.11439868 0.04624819  0.04029308 0.1160492 0.2083065
#> p_3 0.11342316 0.04659492  0.03790970 0.1058081 0.2012647
#> mu  0.06603789 0.54517075 -1.08784005 0.1231664 0.8471849
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>            Mean         SD        2.5%         50%     97.5%
#> p_1  0.09773350 0.06956449  0.01653333  0.07987617 0.2945066
#> p_2  0.09198751 0.04852005  0.02362535  0.08527208 0.2089434
#> p_3  0.09373415 0.04678252  0.02969117  0.08714735 0.2231015
#> mu  -2.19103397 1.42761013 -4.22369935 -2.35417040 0.8828452
#> w_1  0.82835821 0.37848371  0.00000000  1.00000000 1.0000000
#> w_2  0.68656716 0.46562934  0.00000000  1.00000000 1.0000000
#> w_3  0.78358209 0.41334761  0.00000000  1.00000000 1.0000000
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>            Mean         SD         2.5%         50%     97.5%
#> p_1  0.10144394 0.06549089  0.009655962  0.09951111 0.2537204
#> p_2  0.10301439 0.06183202  0.014677379  0.09333534 0.2282552
#> p_3  0.10428387 0.04825738  0.032063702  0.10274114 0.2089005
#> mu  -0.08063454 1.24516180 -2.365158621 -0.04706781 2.4568590
#> w_1  0.61940299 0.48735555  0.000000000  1.00000000 1.0000000
#> w_2  0.67910448 0.46857240  0.000000000  1.00000000 1.0000000
#> w_3  0.58208955 0.49506599  0.000000000  1.00000000 1.0000000
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>            Mean         SD         2.5%         50%     97.5%
#> p_1  0.10037715 0.07238333  0.009448109  0.09142011 0.2942932
#> p_2  0.08905221 0.04935013  0.010584831  0.08252995 0.1907366
#> p_3  0.09329984 0.04535222  0.025731479  0.08336735 0.2087757
#> mu  -0.38973182 0.87973699 -2.633286356 -0.35282543 0.9309299
#> w_1  0.73134328 0.44492424  0.000000000  1.00000000 1.0000000
#> w_2  0.66417910 0.47404879  0.000000000  1.00000000 1.0000000
#> w_3  0.70895522 0.45594860  0.000000000  1.00000000 1.0000000
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>            Mean         SD        2.5%         50%      97.5%
#> p_1  0.09603272 0.05981212  0.01179265  0.09055277  0.2386408
#> p_2  0.09944328 0.05425500  0.02791187  0.09676626  0.2276628
#> p_3  0.09925029 0.04503643  0.03714694  0.09252128  0.2067330
#> mu  -2.43280117 0.66149187 -3.79249505 -2.45567498 -1.2678345
#> w_1  0.69402985 0.46254627  0.00000000  1.00000000  1.0000000
#> w_2  0.76865672 0.42327366  0.00000000  1.00000000  1.0000000
#> w_3  0.87313433 0.33407130  0.00000000  1.00000000  1.0000000
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>     Mean         SD       2.5%        50%     97.5%
#> p_1  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> p_2  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> p_3  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>     Mean         SD        2.5%        50%     97.5%
#> p_1  0.1 0.08660254 0.003720276 0.07636130 0.3240643
#> p_2  0.1 0.06396021 0.014000117 0.08739591 0.2559472
#> p_3  0.1 0.05303301 0.022626655 0.09142896 0.2253392
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>          Mean         SD       2.5%        50%     97.5%
#> p_1 0.1155437 0.09506287 0.00448041 0.09553074 0.2767190
#> p_2 0.1085965 0.06762380 0.01276664 0.09273470 0.2379104
#> p_3 0.1059414 0.05518165 0.02063231 0.09774616 0.2469729
#> 
#> 

## Continuous endpoint example
normal_scenarios <- simulateScenarios(
  n_subjects_list = list(c(30, 30, 30)),
  means_list      = list(c(5.0, 5.5, 6.0)),
  sds_list        = list(c(1, 1, 1)),
  n_trials        = 10,
  endpoint        = "normal")

normal_analyses <- performAnalyses(
  scenario_list      = normal_scenarios,
  method_names       = "normal",
  target_means       = c(5, 5, 5),
  calc_differences   = matrix(c(3, 2, 2, 1), ncol = 2),
  n_mcmc_iterations  = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 12-Jul-2026 Performing Analyses
#>          Analyzing 1 scenario (9 unique trial realizations)
#>    15:54 - with Normal ..........
#>              finished after 0.1 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

getEstimates(normal_analyses)
#> $normal
#> $normal$scenario_1
#>                    Mean        SD        2.5%       50%     97.5%      Bias
#> theta_1       0.1271994 0.2655759 -0.40543262 0.1188432 0.6929612 -4.881157
#> theta_2       0.5079598 0.2648582 -0.02488862 0.5155991 0.9466946 -4.984401
#> theta_3       0.7316392 0.2766652  0.09167425 0.7576424 1.1911010 -5.242358
#> theta_diff_32 0.2236795 0.3701144 -0.52234470 0.2404410 0.8916821  0.740441
#> theta_diff_21 0.3807603 0.3774828 -0.34804868 0.3943990 1.0635905  0.894399
#>                      MSE
#> theta_1       23.8544437
#> theta_2       24.8807628
#> theta_3       27.5342645
#> theta_diff_32  0.6755402
#> theta_diff_21  0.8725109
#> 
#> 
getEstimates(normal_analyses,
             point_estimator = "mean",
             alpha_level     = 0.1)
#> $normal
#> $normal$scenario_1
#>                    Mean        SD          5%       50%       95%       Bias
#> theta_1       0.1271994 0.2655759 -0.27672878 0.1188432 0.5338689 -4.8728006
#> theta_2       0.5079598 0.2648582  0.08187246 0.5155991 0.8801812 -4.9920402
#> theta_3       0.7316392 0.2766652  0.24377477 0.7576424 1.1060765 -5.2683608
#> theta_diff_32 0.2236795 0.3701144 -0.34475809 0.2404410 0.7812519  0.7236795
#> theta_diff_21 0.3807603 0.3774828 -0.21239908 0.3943990 0.9178321  0.8807603
#>                      MSE
#> theta_1       23.7663331
#> theta_2       24.9537403
#> theta_3       27.8025318
#> theta_diff_32  0.6366862
#> theta_diff_21  0.8413100
#> 
#>