scaleRoundList

This function applies scaling and rounding to each item of a list of numerics

scaleRoundList(list, scale_param = 1, round_digits = NULL)

Arguments

list

The list to which the scaling and rounding should be applied to.

scale_param

A numeric for the scaling of each item of the list, Default: `1`

round_digits

An integer for the number of digits. If `NULL`, no rounding will be applied, Default: `NULL`

Value

A list of scaled and rounded numerics

Author

Stephan Wojciekowski

Examples

some_list <- as.list(runif(5))
scaleRoundList(some_list, scale_param = 100, round_digits = 2)
#> [[1]]
#> [1] 31.88
#> 
#> [[2]]
#> [1] 62.31
#> 
#> [[3]]
#> [1] 44.7
#> 
#> [[4]]
#> [1] 70.83
#> 
#> [[5]]
#> [1] 89
#> 

scenarios_list <- simulateScenarios(
  n_subjects_list     = list(c(10, 20, 30)),
  response_rates_list = list(c(0.1, 0.2, 0.3)),
  n_trials            = 10)

analyses_list <- performAnalyses(
  scenario_list       = scenarios_list,
  target_rates        = rep(0.5, 3),
  n_mcmc_iterations   = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 12-Jul-2026 Performing Analyses
#>          Analyzing 1 scenario (10 unique trial realizations)
#>    15:55 - with Berry ...........
#>              finished after 0.1 secs.
#>    15:55 - with Exnex ...........
#>              finished after 0.1 secs.
#>    15:55 - with Exnex_adj .......
#>              finished after 0.1 secs.
#>    15:55 - with Exnex_adj_mix ...
#>              finished after 0.1 secs.
#>    15:55 - with Exnex_mix .......
#>              finished after 0.1 secs.
#>    15:55 - with Pooled ..........
#>              finished after 0 secs.
#>    15:55 - with Stratified ......
#>              finished after 0.1 secs.
#>    15:55 - with Stratified_mix ...
#>              finished after 0.1 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

scaleRoundList(
  list         = getEstimates(analyses_list),
  scale_param  = 100,
  round_digits = 2)
#> $berry
#> $berry$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 15.33 8.57  2.53 14.75 32.44  4.75 0.32
#> p_2 21.94 7.27  9.69 21.53 37.47  1.53 0.24
#> p_3 26.37 6.98 14.99 25.79 41.39 -4.21 0.76
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 12.51 8.61  1.83 10.56 32.81  0.56 0.29
#> p_2 22.77 8.17  8.87 22.41 39.43  2.41 0.44
#> p_3 28.47 7.94 15.55 27.91 45.82 -2.09 0.72
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 13.08 8.50  2.00 11.39 32.40  1.39 0.32
#> p_2 22.93 8.13  9.47 22.08 40.23  2.08 0.34
#> p_3 28.70 7.63 16.32 28.26 45.07 -1.74 0.75
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 12.26 8.43  1.58 10.54 32.31  0.54 0.36
#> p_2 22.80 8.40  9.14 22.00 39.69  2.00 0.36
#> p_3 28.85 7.87 14.83 28.54 44.83 -1.46 0.81
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 12.96 8.65  1.70 11.25 33.91  1.25 0.29
#> p_2 22.54 8.41  8.83 21.65 39.69  1.65 0.33
#> p_3 28.59 7.74 15.61 28.09 44.12 -1.91 0.70
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 22.95 5.29 13.48 22.65  34.1 12.65 1.85
#> p_2 22.95 5.29 13.48 22.65  34.1  2.65 0.32
#> p_3 22.95 5.29 13.48 22.65  34.1 -7.35 0.79
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1  9.09 7.79  0.77  6.85 29.50 -3.15 0.54
#> p_2 22.86 8.82  8.31 21.99 42.26  1.99 0.47
#> p_3 29.68 7.91 15.56 29.24 46.27 -0.76 0.80
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1  9.58 8.03  1.19  7.55 31.18 -2.45 0.53
#> p_2 23.15 8.83  8.92 22.70 41.62  2.70 0.53
#> p_3 29.84 7.90 15.58 29.12 45.39 -0.88 0.88
#> 
#>