setPriorParametersExNexAdj

This function sets prior parameters for the analysis method "exnex_adj" for use in performAnalyses.

setPriorParametersExNexAdj(mu_mean, mu_sd, tau_scale, mu_j, tau_j, w_j)

Arguments

mu_mean

A numeric for the mean of \(\mu\)

mu_sd

A positive numeric for the standard deviation of \(\mu\)

tau_scale

A positive numeric for the scale parameter of \(\tau\)

mu_j

A vector of numerics for the means \(\mu_j\)

tau_j

A vector of positive numerics for the standard deviations \(\tau_j\)

w_j

A numeric in (0, 1) for the weight of the Ex component

Value

A list with prior parameters of class prior_parameters_list

Details

This function sets the prior parameters for the method ExNex Adjusted, which combines the approach proposed by Neuenschwander et al. (2016) and the approach proposed by Berry et al. (2013).

See also

Author

Stephan Wojciekowski

Examples

 prior_parameters_exnex_adj <- setPriorParametersExNexAdj(0, 1, 2, c(4, 5), c(6, 7), 0.8)