setPriorParametersNormal

This function sets prior parameters for the analysis method "normal" for use in performAnalyses.

setPriorParametersNormal(
  mu_mean,
  mu_sd,
  tau_scale,
  mu_j,
  tau_j,
  w_j,
  target_means,
  sigma_shape,
  sigma_rate
)

Arguments

mu_mean

A numeric for the mean of the Ex component.

mu_sd

A positive numeric for the standard deviation of the Ex component.

tau_scale

A positive numeric for the scale parameter of the Half-normal prior on the between-cohort standard deviation \(\tau\).

mu_j

A numeric vector for the means of the Nex components.

tau_j

A positive numeric vector for the standard deviations of the Nex components.

w_j

A numeric in (0, 1) for the weight of the Ex component.

target_means

A numeric vector of target means for each cohort. These are used as additive offsets in the adjusted normal ExNex model.

sigma_shape

A positive numeric for the Gamma shape prior of the residual precision.

sigma_rate

A positive numeric for the Gamma rate prior of the residual precision.

Value

A list with prior parameters of class prior_parameters_list

Details

The method "normal" is an adjusted ExNex hierarchical normal model for continuous endpoints. The prior parameters specify the Ex component, the cohort-specific Nex components, the Ex/Nex mixture weight, the target means used for adjustment, and the residual precision prior. In applications with continuous endpoints on very different scales, the residual precision prior should be chosen with regard to the endpoint scale.

See also

Author

Stephan Wojciekowski

Examples

prior_parameters_normal <- setPriorParametersNormal(
  mu_mean      = 0,
  mu_sd        = 1,
  tau_scale    = 1,
  mu_j         = c(0, 0, 0),
  tau_j        = c(1, 1, 1),
  w_j          = 0.5,
  target_means = c(5, 5, 5),
  sigma_shape  = 1,
  sigma_rate   = 1
)