getEstimates

This function calculates the point estimates and credible intervals per cohort, as well as estimates of the biases and the mean squared errors of the point estimates per cohort.

getEstimates(
  analyses_list,
  add_parameters = NULL,
  point_estimator = "median",
  alpha_level = 0.05
)

Arguments

analyses_list

An object of class analysis_list, as created with performAnalyses

add_parameters

A vector of strings naming additional parameters from the Bayesian hierarchical models, e.g. c('mu', 'tau'). If NULL, no additional parameters will be evaluated, Default: NULL

point_estimator

A string indicating the type of estimator used for calculation of bias and MSE. Must be one of 'median' or 'mean'

alpha_level

A numeric in (0, 1) for the level of the credible interval. Only values corresponding to quantiles saved in performAnalyses will work, Default: 0.05

Value

A named list of matrices of estimates of cohort-level parameters and credible intervals. Estimates of bias and MSE are included for cohort-level posterior estimates of simulated trials.

Details

For binary endpoints, bias and MSE will only be calculated for posterior response-rate estimates of simulated trials. For continuous endpoints, bias and MSE will only be calculated for posterior cohort-mean estimates of simulated trials. For additional parameters, bias and MSE will not be calculated.

Possible additional parameters for the Bayesian hierarchical models are c('mu', 'tau') for 'berry', 'exnex', 'exnex_mix', 'exnex_adj', and 'exnex_adj_mix'. The ExNex-based models can also access posterior weights. paste0("w_", seq_len(n_cohorts)).

See also

Author

Stephan Wojciekowski

Examples

  scenarios_list <- simulateScenarios(
    n_subjects_list     = list(c(10, 20, 30)),
    response_rates_list = list(c(0.1, 0.2, 3)),
    n_trials            = 10)

  analyses_list <- performAnalyses(
    scenario_list       = scenarios_list,
    target_rates        = c(0.1, 0.1, 0.1),
    calc_differences    = matrix(c(3, 2, 2, 1), ncol = 2),
    n_mcmc_iterations   = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 24-Apr-2026 Performing Analyses
#>          Analyzing 1 scenario (9 unique trial realizations)
#>    12:43 - with Berry ...........
#>              finished after 0.1 secs.
#>    12:43 - with Exnex ...........
#>              finished after 0.1 secs.
#>    12:43 - with Exnex_adj .......
#>              finished after 0.1 secs.
#>    12:43 - with Exnex_adj_mix ...
#>              finished after 0.1 secs.
#>    12:43 - with Exnex_mix .......
#>              finished after 0.1 secs.
#>    12:43 - with Pooled ..........
#>              finished after 0.1 secs.
#>    12:43 - with Stratified ......
#>              finished after 0.1 secs.
#>    12:43 - with Stratified_mix ...
#>              finished after 0.1 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

  getEstimates(analyses_list)
#> $berry
#> $berry$scenario_1
#>                  Mean         SD        2.5%         50%      97.5%        Bias
#> p_1        0.14227305 0.07214596  0.03238457  0.13443249 0.30937244  0.03443249
#> p_2        0.21929589 0.07578783  0.09667009  0.21160328 0.37817724  0.01160328
#> p_3        0.12495036 0.05305730  0.04038027  0.12117943 0.24130048  0.02117943
#> p_diff_32 -0.09434554 0.08466067 -0.26465547 -0.08996266 0.05818028 -0.18996266
#> p_diff_21  0.07702284 0.09462066 -0.10418235  0.07231066 0.26501352  0.17231066
#>                    MSE
#> p_1       0.0038577129
#> p_2       0.0066963381
#> p_3       0.0008388088
#> p_diff_32 0.0434368829
#> p_diff_21 0.0365939218
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>                 Mean         SD        2.5%         50%      97.5%         Bias
#> p_1        0.1423861 0.08452425  0.02733764  0.12730792 0.33634287  0.027307918
#> p_2        0.2306020 0.08487120  0.09518284  0.22017216 0.40614967  0.020172159
#> p_3        0.1117141 0.05399676  0.02813088  0.10334476 0.23070042  0.003344757
#> p_diff_32 -0.1188879 0.10193980 -0.33079508 -0.11186780 0.06631459 -0.211867796
#> p_diff_21  0.0882159 0.11884091 -0.13600876  0.08689998 0.31910373  0.186899980
#>                    MSE
#> p_1       0.0059799113
#> p_2       0.0096210872
#> p_3       0.0001829617
#> p_diff_32 0.0548504657
#> p_diff_21 0.0451611978
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>                 Mean         SD        2.5%         50%      97.5%         Bias
#> p_1        0.1384767 0.08534402  0.02285198  0.12171728 0.33930515  0.021717276
#> p_2        0.2333182 0.08128531  0.10833972  0.22199666 0.40774037  0.021996661
#> p_3        0.1119843 0.05456824  0.02917809  0.10522959 0.22995166  0.005229592
#> p_diff_32 -0.1213338 0.09435459 -0.31750254 -0.11200569 0.03234662 -0.212005689
#> p_diff_21  0.0948415 0.11462290 -0.14337347  0.09632857 0.31212079  0.196328570
#>                    MSE
#> p_1       0.0062605608
#> p_2       0.0097929641
#> p_3       0.0001071203
#> p_diff_32 0.0550316899
#> p_diff_21 0.0511139531
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>                  Mean         SD        2.5%         50%      97.5%
#> p_1        0.13589569 0.08404904  0.02170071  0.12273804 0.33088492
#> p_2        0.23438747 0.08366122  0.10379327  0.22584983 0.42763679
#> p_3        0.11400117 0.05644309  0.02905822  0.10626195 0.23845485
#> p_diff_32 -0.12038630 0.10074211 -0.33576841 -0.11000659 0.04389818
#> p_diff_21  0.09849178 0.11653452 -0.12100653  0.09246087 0.33031423
#>                   Bias          MSE
#> p_1        0.022738036 7.930390e-03
#> p_2        0.025849829 9.097584e-03
#> p_3        0.006261945 8.596061e-05
#> p_diff_32 -0.210006592 5.346830e-02
#> p_diff_21  0.192460874 4.631276e-02
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>                 Mean         SD        2.5%        50%      97.5%          Bias
#> p_1        0.1367098 0.08704136  0.02542711  0.1187758 0.33973899  0.0187758348
#> p_2        0.2434169 0.08781298  0.09774983  0.2382065 0.42279351  0.0382064735
#> p_3        0.1086583 0.05111788  0.03138670  0.1008468 0.22782660  0.0008467725
#> p_diff_32 -0.1347587 0.09913597 -0.33949710 -0.1284773 0.03681498 -0.2284772548
#> p_diff_21  0.1067071 0.11991615 -0.14198138  0.1119493 0.33283414  0.2119493165
#>                    MSE
#> p_1       5.887833e-03
#> p_2       1.174380e-02
#> p_3       9.201092e-05
#> p_diff_32 6.340072e-02
#> p_diff_21 5.859824e-02
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>               Mean         SD       2.5%       50%     97.5%        Bias
#> p_1       0.157377 0.04585289 0.07888073 0.1536238 0.2570606  0.05362382
#> p_2       0.157377 0.04585289 0.07888073 0.1536238 0.2570606 -0.04637618
#> p_3       0.157377 0.04585289 0.07888073 0.1536238 0.2570606  0.05362382
#> p_diff_32 0.000000 0.00000000 0.00000000 0.0000000 0.0000000 -0.10000000
#> p_diff_21 0.000000 0.00000000 0.00000000 0.0000000 0.0000000  0.10000000
#>                   MSE
#> p_1       0.004294231
#> p_2       0.003569467
#> p_3       0.004294231
#> p_diff_32 0.010000000
#> p_diff_21 0.010000000
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>                 Mean         SD        2.5%         50%      97.5%         Bias
#> p_1        0.1272727 0.08714414  0.01724810  0.10894029 0.34120549  0.008940287
#> p_2        0.2523810 0.08983870  0.10146187  0.24443869 0.44755570  0.044438692
#> p_3        0.1000000 0.05303301  0.02262665  0.09142896 0.22533918 -0.008571039
#> p_diff_32 -0.1579533 0.10882163 -0.37989075 -0.15244793 0.03307176 -0.252447934
#> p_diff_21  0.1203415 0.12815262 -0.12764464  0.12531731 0.34922036  0.225317306
#>                    MSE
#> p_1       7.325412e-03
#> p_2       1.205897e-02
#> p_3       7.346271e-05
#> p_diff_32 7.312230e-02
#> p_diff_21 6.569212e-02
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>                 Mean         SD        2.5%         50%     97.5%          Bias
#> p_1        0.1424413 0.09359026  0.02610316  0.11955607 0.3702330  0.0195560714
#> p_2        0.2638824 0.09153531  0.11626568  0.25965652 0.4577202  0.0596565194
#> p_3        0.1059414 0.05518165  0.02980409  0.09985911 0.2335491 -0.0001408894
#> p_diff_32 -0.1559290 0.11288950 -0.36122771 -0.15240504 0.0532696 -0.2524050379
#> p_diff_21  0.1216104 0.13962052 -0.14439135  0.12971567 0.3680204  0.2297156735
#>                    MSE
#> p_1       9.021016e-03
#> p_2       1.357342e-02
#> p_3       3.521128e-05
#> p_diff_32 7.283963e-02
#> p_diff_21 6.838489e-02
#> 
#> 
  getEstimates(analyses_list   = analyses_list,
               add_parameters  = c("mu", "tau", "w_1", "w_2", "w_3"),
               point_estimator = "mean",
               alpha_level     = 0.1)
#> $berry
#> $berry$scenario_1
#>                  Mean         SD          5%         50%        95%        Bias
#> p_1        0.14227305 0.07214596  0.04159033  0.13443249 0.27092724  0.04227305
#> p_2        0.21929589 0.07578783  0.10917439  0.21160328 0.35553850  0.01929589
#> p_3        0.12495036 0.05305730  0.04905488  0.12117943 0.21821221  0.02495036
#> p_diff_32 -0.09434554 0.08466067 -0.23467119 -0.08996266 0.02887673 -0.19434554
#> p_diff_21  0.07702284 0.09462066 -0.07285345  0.07231066 0.22975114  0.17702284
#> mu         0.34915828 0.65918467 -0.83597093  0.41449770 1.28062104          NA
#> tau        0.73116788 0.47515266  0.15142278  0.61681681 1.61357796          NA
#>                   MSE
#> p_1       0.004268074
#> p_2       0.007654453
#> p_3       0.000824041
#> p_diff_32 0.044959009
#> p_diff_21 0.038574302
#> mu                 NA
#> tau                NA
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>                 Mean         SD          5%         50%         95%        Bias
#> p_1        0.1423861 0.08452425  0.03582518  0.12730792  0.30808265  0.04238605
#> p_2        0.2306020 0.08487120  0.11007226  0.22017216  0.37997746  0.03060195
#> p_3        0.1117141 0.05399676  0.03730535  0.10334476  0.21185781  0.01171407
#> p_diff_32 -0.1188879 0.10193980 -0.29043635 -0.11186780  0.03117522 -0.21888788
#> p_diff_21  0.0882159 0.11884091 -0.11295775  0.08689998  0.28081662  0.18821590
#> mu        -1.9858044 1.36520383 -4.30958228 -1.81219726 -0.23930087          NA
#> tau        0.5933130 0.46202710  0.07261477  0.49873749  1.48083983          NA
#> w_1        0.6335821 0.47304179  0.00000000  0.90000000  1.00000000          NA
#> w_2        0.6276119 0.47059263  0.00000000  0.80000000  1.00000000          NA
#> w_3        0.6074627 0.46488491  0.00000000  0.80000000  1.00000000          NA
#>                    MSE
#> p_1       0.0068045242
#> p_2       0.0099284219
#> p_3       0.0002547294
#> p_diff_32 0.0568339315
#> p_diff_21 0.0467928794
#> mu                  NA
#> tau                 NA
#> w_1                 NA
#> w_2                 NA
#> w_3                 NA
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>                 Mean         SD          5%         50%        95%        Bias
#> p_1        0.1384767 0.08534402  0.03300132  0.12171728 0.30104803  0.03847666
#> p_2        0.2333182 0.08128531  0.12143567  0.22199666 0.37249139  0.03331816
#> p_3        0.1119843 0.05456824  0.03736009  0.10522959 0.21074670  0.01198434
#> p_diff_32 -0.1213338 0.09435459 -0.28916186 -0.11200569 0.01330783 -0.22133383
#> p_diff_21  0.0948415 0.11462290 -0.09650906  0.09632857 0.28035819  0.19484150
#> mu         0.2605755 1.24808388 -2.02807355  0.41144279 1.88094572          NA
#> tau        0.6865201 0.50096531  0.12315599  0.57123913 1.64068365          NA
#> w_1        0.5977612 0.46836600  0.00000000  0.70000000 1.00000000          NA
#> w_2        0.6455224 0.46245624  0.00000000  0.80000000 1.00000000          NA
#> w_3        0.6223881 0.46331190  0.00000000  0.80000000 1.00000000          NA
#>                    MSE
#> p_1       0.0069927362
#> p_2       0.0101816242
#> p_3       0.0002025248
#> p_diff_32 0.0583441579
#> p_diff_21 0.0511468582
#> mu                  NA
#> tau                 NA
#> w_1                 NA
#> w_2                 NA
#> w_3                 NA
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>                  Mean         SD          5%         50%        95%        Bias
#> p_1        0.13589569 0.08404904  0.02970566  0.12273804 0.29088636  0.03589569
#> p_2        0.23438747 0.08366122  0.11667928  0.22584983 0.38192243  0.03438747
#> p_3        0.11400117 0.05644309  0.03785834  0.10626195 0.21470489  0.01400117
#> p_diff_32 -0.12038630 0.10074211 -0.29482586 -0.11000659 0.02218757 -0.22038630
#> p_diff_21  0.09849178 0.11653452 -0.07641335  0.09246087 0.28231623  0.19849178
#> mu         0.35024229 1.16706515 -1.81504543  0.53174536 1.78232229          NA
#> tau        0.59971604 0.48265334  0.06279314  0.47356012 1.47602133          NA
#> w_1        0.61865672 0.44581711  0.00000000  0.80000000 1.00000000          NA
#> w_2        0.74925373 0.41396800  0.00000000  1.00000000 1.00000000          NA
#> w_3        0.62611940 0.45447310  0.00000000  0.80000000 1.00000000          NA
#>                    MSE
#> p_1       0.0076991003
#> p_2       0.0091184302
#> p_3       0.0002375155
#> p_diff_32 0.0563270296
#> p_diff_21 0.0486863705
#> mu                  NA
#> tau                 NA
#> w_1                 NA
#> w_2                 NA
#> w_3                 NA
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>                 Mean         SD          5%        50%          95%
#> p_1        0.1367098 0.08704136  0.03270187  0.1187758  0.303516652
#> p_2        0.2434169 0.08781298  0.11576261  0.2382065  0.393377587
#> p_3        0.1086583 0.05111788  0.03897731  0.1008468  0.206637443
#> p_diff_32 -0.1347587 0.09913597 -0.30032934 -0.1284773  0.009252947
#> p_diff_21  0.1067071 0.11991615 -0.09881464  0.1119493  0.308142889
#> mu        -1.8518133 1.25975108 -3.65521311 -1.8404818 -0.093767920
#> tau        0.6976970 0.49434071  0.11030487  0.5731828  1.662478607
#> w_1        0.6007463 0.47050368  0.00000000  0.7000000  1.000000000
#> w_2        0.6208955 0.44681802  0.00000000  0.8000000  1.000000000
#> w_3        0.6574627 0.45322961  0.00000000  0.9000000  1.000000000
#>                   Bias         MSE
#> p_1        0.036709821 0.006392884
#> p_2        0.043416939 0.012408358
#> p_3        0.008658265 0.000119273
#> p_diff_32 -0.234758674 0.066387734
#> p_diff_21  0.206707118 0.057811391
#> mu                  NA          NA
#> tau                 NA          NA
#> w_1                 NA          NA
#> w_2                 NA          NA
#> w_3                 NA          NA
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>               Mean         SD         5%       50%       95%        Bias
#> p_1       0.157377 0.04585289 0.08883602 0.1536238 0.2387526  0.05737705
#> p_2       0.157377 0.04585289 0.08883602 0.1536238 0.2387526 -0.04262295
#> p_3       0.157377 0.04585289 0.08883602 0.1536238 0.2387526  0.05737705
#> p_diff_32 0.000000 0.00000000 0.00000000 0.0000000 0.0000000 -0.10000000
#> p_diff_21 0.000000 0.00000000 0.00000000 0.0000000 0.0000000  0.10000000
#>                   MSE
#> p_1       0.004680641
#> p_2       0.003205232
#> p_3       0.004680641
#> p_diff_32 0.010000000
#> p_diff_21 0.010000000
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>                 Mean         SD          5%         50%         95%        Bias
#> p_1        0.1272727 0.08714414  0.02447923  0.10894029 0.294035673  0.02727273
#> p_2        0.2523810 0.08983870  0.11927312  0.24443869 0.412744522  0.05238095
#> p_3        0.1000000 0.05303301  0.02959876  0.09142896 0.199781011  0.00000000
#> p_diff_32 -0.1579533 0.10882163 -0.34167308 -0.15244793 0.005204592 -0.25795325
#> p_diff_21  0.1203415 0.12815262 -0.08922427  0.12531731 0.313878596  0.22034151
#>                   MSE
#> p_1       0.008181818
#> p_2       0.012217183
#> p_3       0.000000000
#> p_diff_32 0.075814200
#> p_diff_21 0.064512196
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>                 Mean         SD          5%         50%        95%         Bias
#> p_1        0.1424413 0.09359026  0.03349981  0.11955607 0.31448820  0.042441309
#> p_2        0.2638824 0.09153531  0.13882943  0.25965652 0.41276493  0.063882389
#> p_3        0.1059414 0.05518165  0.03604901  0.09985911 0.21199758  0.005941434
#> p_diff_32 -0.1559290 0.11288950 -0.33418832 -0.15240504 0.02053212 -0.255929030
#> p_diff_21  0.1216104 0.13962052 -0.11116242  0.12971567 0.33909972  0.221610433
#>                    MSE
#> p_1       1.045677e-02
#> p_2       1.364609e-02
#> p_3       3.530063e-05
#> p_diff_32 7.339775e-02
#> p_diff_21 6.661158e-02
#> 
#> 

  outcome <- createTrial(
    n_subjects          = c(10, 20, 30),
    n_responders        = c( 1,  2,  3))

  outcome_analysis <- performAnalyses(
    scenario_list       = outcome,
    target_rates        = c(0.1, 0.1, 0.1),
    n_mcmc_iterations   = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 24-Apr-2026 Performing Analyses
#>          Analyzing 1 scenario (1 unique trial realization)
#>    12:43 - with Berry ...........
#>              finished after 0.1 secs.
#>    12:43 - with Exnex ...........
#>              finished after 0.1 secs.
#>    12:43 - with Exnex_adj .......
#>              finished after 0.1 secs.
#>    12:43 - with Exnex_adj_mix ...
#>              finished after 0.1 secs.
#>    12:43 - with Exnex_mix .......
#>              finished after 0.1 secs.
#>    12:43 - with Pooled ..........
#>              finished after 0.1 secs.
#>    12:43 - with Stratified ......
#>              finished after 0.1 secs.
#>    12:43 - with Stratified_mix ...
#>              finished after 0.1 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

  getEstimates(outcome_analysis)
#> $berry
#> $berry$scenario_1
#>           Mean         SD       2.5%        50%     97.5%
#> p_1 0.10305867 0.06430017 0.02800930 0.08704086 0.2413489
#> p_2 0.09661951 0.04722316 0.02964679 0.08523771 0.2137815
#> p_3 0.09169210 0.04050202 0.03317640 0.08686999 0.1855336
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>           Mean         SD        2.5%        50%     97.5%
#> p_1 0.10218945 0.06801584 0.008273141 0.08919231 0.2827500
#> p_2 0.10161309 0.05983160 0.018560624 0.09270514 0.2307385
#> p_3 0.09409026 0.04210250 0.024433561 0.09374918 0.1925546
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>           Mean         SD       2.5%        50%     97.5%
#> p_1 0.10356685 0.07415484 0.01514351 0.08112209 0.2493845
#> p_2 0.09952007 0.05840382 0.01704206 0.09132356 0.2396303
#> p_3 0.09656930 0.04582350 0.03738565 0.08638407 0.1983649
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>           Mean         SD       2.5%        50%     97.5%
#> p_1 0.10032992 0.06457458 0.01640226 0.08947104 0.2514302
#> p_2 0.09684415 0.04578778 0.02052813 0.09182585 0.2001136
#> p_3 0.09547945 0.04106454 0.03014167 0.09120540 0.1736810
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>           Mean         SD        2.5%        50%     97.5%
#> p_1 0.09506157 0.06788035 0.005869617 0.07664532 0.2540066
#> p_2 0.09864069 0.05719578 0.031841144 0.08465303 0.2269160
#> p_3 0.08249722 0.03972890 0.019603061 0.07884341 0.1648669
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>     Mean         SD       2.5%        50%     97.5%
#> p_1  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> p_2  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> p_3  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>     Mean         SD        2.5%        50%     97.5%
#> p_1  0.1 0.08660254 0.003720276 0.07636130 0.3240643
#> p_2  0.1 0.06396021 0.014000117 0.08739591 0.2559472
#> p_3  0.1 0.05303301 0.022626655 0.09142896 0.2253392
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>          Mean         SD        2.5%        50%     97.5%
#> p_1 0.1155437 0.09506287 0.003415863 0.10360886 0.3934643
#> p_2 0.1085965 0.06762380 0.020649613 0.10019381 0.2758158
#> p_3 0.1059414 0.05518165 0.019049949 0.09419874 0.2186897
#> 
#> 
  getEstimates(analyses_list  = outcome_analysis,
               add_parameters = c("mu", "w_1", "w_2", "w_3"))
#> $berry
#> $berry$scenario_1
#>            Mean         SD        2.5%         50%     97.5%
#> p_1  0.10305867 0.06430017  0.02800930  0.08704086 0.2413489
#> p_2  0.09661951 0.04722316  0.02964679  0.08523771 0.2137815
#> p_3  0.09169210 0.04050202  0.03317640  0.08686999 0.1855336
#> mu  -0.20028133 0.57779142 -1.34579182 -0.21939129 0.8632300
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>            Mean         SD         2.5%         50%      97.5%
#> p_1  0.10218945 0.06801584  0.008273141  0.08919231  0.2827500
#> p_2  0.10161309 0.05983160  0.018560624  0.09270514  0.2307385
#> p_3  0.09409026 0.04210250  0.024433561  0.09374918  0.1925546
#> mu  -2.35135581 0.52781738 -3.448480694 -2.24346427 -1.5359052
#> w_1  0.73134328 0.44492424  0.000000000  1.00000000  1.0000000
#> w_2  0.79104478 0.40808791  0.000000000  1.00000000  1.0000000
#> w_3  0.85820896 0.35014465  0.000000000  1.00000000  1.0000000
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>            Mean         SD        2.5%         50%     97.5%
#> p_1  0.10356685 0.07415484  0.01514351  0.08112209 0.2493845
#> p_2  0.09952007 0.05840382  0.01704206  0.09132356 0.2396303
#> p_3  0.09656930 0.04582350  0.03738565  0.08638407 0.1983649
#> mu  -0.12707749 0.69256924 -1.18103174 -0.20156996 1.1526338
#> w_1  0.74626866 0.43677849  0.00000000  1.00000000 1.0000000
#> w_2  0.79104478 0.40808791  0.00000000  1.00000000 1.0000000
#> w_3  0.72388060 0.44875418  0.00000000  1.00000000 1.0000000
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>           Mean         SD        2.5%        50%     97.5%
#> p_1 0.10032992 0.06457458  0.01640226 0.08947104 0.2514302
#> p_2 0.09684415 0.04578778  0.02052813 0.09182585 0.2001136
#> p_3 0.09547945 0.04106454  0.03014167 0.09120540 0.1736810
#> mu  0.11215381 0.81756482 -1.23314507 0.05119025 2.4202104
#> w_1 0.80597015 0.39693586  0.00000000 1.00000000 1.0000000
#> w_2 0.74626866 0.43677849  0.00000000 1.00000000 1.0000000
#> w_3 0.80597015 0.39693586  0.00000000 1.00000000 1.0000000
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>            Mean         SD         2.5%         50%     97.5%
#> p_1  0.09506157 0.06788035  0.005869617  0.07664532 0.2540066
#> p_2  0.09864069 0.05719578  0.031841144  0.08465303 0.2269160
#> p_3  0.08249722 0.03972890  0.019603061  0.07884341 0.1648669
#> mu  -2.45105530 1.47026132 -4.600887167 -2.43842383 0.4772009
#> w_1  0.58208955 0.49506599  0.000000000  1.00000000 1.0000000
#> w_2  0.56716418 0.49732761  0.000000000  1.00000000 1.0000000
#> w_3  0.51492537 0.50165253  0.000000000  1.00000000 1.0000000
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>     Mean         SD       2.5%        50%     97.5%
#> p_1  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> p_2  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> p_3  0.1 0.03810004 0.03862692 0.09563321 0.1860054
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>     Mean         SD        2.5%        50%     97.5%
#> p_1  0.1 0.08660254 0.003720276 0.07636130 0.3240643
#> p_2  0.1 0.06396021 0.014000117 0.08739591 0.2559472
#> p_3  0.1 0.05303301 0.022626655 0.09142896 0.2253392
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>          Mean         SD        2.5%        50%     97.5%
#> p_1 0.1155437 0.09506287 0.003415863 0.10360886 0.3934643
#> p_2 0.1085965 0.06762380 0.020649613 0.10019381 0.2758158
#> p_3 0.1059414 0.05518165 0.019049949 0.09419874 0.2186897
#> 
#> 

  normal_scenarios <- simulateScenarios(
    n_subjects_list = list(c(30, 30, 30)),
    means_list      = list(c(5.0, 5.5, 6.0)),
    sds_list        = list(c(1, 1, 1)),
    n_trials        = 10,
    endpoint        = "normal")

  normal_analyses <- performAnalyses(
    scenario_list      = normal_scenarios,
    method_names       = "normal",
    target_means       = c(5, 5, 5),
    calc_differences   = matrix(c(3, 2, 2, 1), ncol = 2),
    n_mcmc_iterations  = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 24-Apr-2026 Performing Analyses
#>          Analyzing 1 scenario (10 unique trial realizations)
#>    12:43 - with Normal ..........
#>              finished after 0.1 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

  getEstimates(normal_analyses)
#> $normal
#> $normal$scenario_1
#>                     Mean        SD       2.5%        50%     97.5%       Bias
#> theta_1       0.02077677 0.2816121 -0.5386953 0.01485436 0.6211321 -4.9851456
#> theta_2       0.50150221 0.2640445 -0.0182997 0.49997893 0.9915045 -5.0000211
#> theta_3       0.93740100 0.2748125  0.3155570 0.96789011 1.4051987 -5.0321099
#> theta_diff_32 0.43589879 0.3668223 -0.3347986 0.45045509 1.1122251  0.9504551
#> theta_diff_21 0.48072544 0.3817088 -0.2751629 0.49367727 1.2171995  0.9936773
#>                      MSE
#> theta_1       24.8906759
#> theta_2       25.0350322
#> theta_3       25.3406757
#> theta_diff_32  0.9355242
#> theta_diff_21  1.0474098
#> 
#> 
  getEstimates(analyses_list   = normal_analyses,
               point_estimator = "mean",
               alpha_level     = 0.1)
#> $normal
#> $normal$scenario_1
#>                     Mean        SD          5%        50%       95%       Bias
#> theta_1       0.02077677 0.2816121 -0.39686618 0.01485436 0.4546015 -4.9792232
#> theta_2       0.50150221 0.2640445  0.09133971 0.49997893 0.8969610 -4.9984978
#> theta_3       0.93740100 0.2748125  0.44762581 0.96789011 1.2908104 -5.0625990
#> theta_diff_32 0.43589879 0.3668223 -0.13279667 0.45045509 0.9725329  0.9358988
#> theta_diff_21 0.48072544 0.3817088 -0.08768963 0.49367727 1.0511720  0.9807254
#>                      MSE
#> theta_1       24.8270133
#> theta_2       25.0152587
#> theta_3       25.6487847
#> theta_diff_32  0.9062791
#> theta_diff_21  1.0179519
#> 
#>