getPriorParameters

This function provides default prior parameters for the analysis methods that can be used in performAnalyses.

getPriorParameters(
  method_names,
  target_rates,
  n_worth = 1,
  tau_scale = 1,
  w_j = 0.5
)

Arguments

method_names

A vector of strings for the names of the methods to be used. Available methods: c("berry", "exnex", "exnex_adj", "pooled", "stratified")

target_rates

A vector of numerics in (0, 1) for the target rate of each cohort

n_worth

An integer for the number of subjects the variability of the prior should reflect response rate scale, Default: 1

tau_scale

A numeric for the scale parameter of the Half-normal distribution of \(\tau\) in the methods "berry", "exnex", and "exnex_adj", Default: 1

w_j

A numeric in (0, 1) for the weight of the Ex component in the methods "exnex" and "exnex_adj", Default: 0.5

Value

A list with prior parameters of class prior_parameters_list

Details

Regarding the default prior parameters for "berry", "exnex", and "exnex_adj":

"berry": The mean of \(\mu\) is set to 0. Its variance is calculated as proposed in "Robust exchangeability designs for early phase clinical trials with multiple strata" (Neuenschwander et al. (2016)) with regard to n_worth. The scale parameter of \(\tau\) is set to tau_scale.

"exnex": The weight of the Ex component is set to w_j. For the Ex component: The target rate that results in the greatest variance is determined. The mean of \(\mu\) is set to that target rate. The variance of \(\mu\) is calculated as proposed in "Robust exchangeability designs for early phase clinical trials with multiple strata" (Neuenschwander et al. (2016)) with regard to n_worth. The scale parameter of \(\tau\) is set to tau_scale. For the Nex components: The means of \(\mu_j\) are set to the respective target rates. The variances of \(\tau_j\) are calculated as proposed in "Robust exchangeability designs for early phase clinical trials with multiple strata" (Neuenschwander et al. (2016)) with regard to n_worth, see also getMuVar.

"exnex_adj": The weight of the Ex component is set to w_j. For the Ex component: The target rate that results in the greatest variance is determined. The mean of \(\mu\) is set to 0. The variance of \(\mu\) is calculated as proposed in "Robust exchangeability designs for early phase clinical trials with multiple strata" (Neuenschwander et al. (2016)) with regard to n_worth, see also getMuVar. The scale parameter of \(\tau\) is set to tau_scale. For the Nex components: The means of \(\mu_j\) are set to the 0. The variances of \(\tau_j\) are calculated as proposed in "Robust exchangeability designs for early phase clinical trials with multiple strata" (Neuenschwander et al. (2016)) with regard to n_worth, see also getMuVar.

"pooled": The target rate that results in the greatest variance is determined. The scale parameter \(\alpha\) is set to that target rate times n_worth. The scale parameter \(\beta\) is set to 1 - that target rate times n_worth.

"stratified": The scale parameters \(\alpha_j\) are set to target_rates * n_worth. The scale parameters \(\beta_j\) are set to (1 - target_rates) * n_worth.

References

Berry, Scott M., et al. "Bayesian hierarchical modeling of patient subpopulations: efficient designs of phase II oncology clinical trials." Clinical Trials 10.5 (2013): 720-734.

Neuenschwander, Beat, et al. "Robust exchangeability designs for early phase clinical trials with multiple strata." Pharmaceutical statistics 15.2 (2016): 123-134.

See also

Author

Stephan Wojciekowski

Examples

prior_parameters_list <- getPriorParameters(
  method_names = c("berry", "exnex", "exnex_adj", "pooled", "stratified"),
  target_rates = c(0.1, 0.2, 0.3))