scaleRoundList

This function applies scaling and rounding to each item of a list of numerics

scaleRoundList(list, scale_param = 1, round_digits = NULL)

Arguments

list

The list to which the scaling and rounding should be applied to.

scale_param

A numeric for the scaling of each item of the list, Default: `1`

round_digits

An integer for the number of digits. If `NULL`, no rounding will be applied, Default: `NULL`

Value

A list of scaled and rounded numerics

Author

Stephan Wojciekowski

Examples

some_list <- as.list(runif(5))
scaleRoundList(some_list, scale_param = 100, round_digits = 2)
#> [[1]]
#> [1] 55.29
#> 
#> [[2]]
#> [1] 21.92
#> 
#> [[3]]
#> [1] 64.67
#> 
#> [[4]]
#> [1] 72.77
#> 
#> [[5]]
#> [1] 87.26
#> 

scenarios_list <- simulateScenarios(
  n_subjects_list     = list(c(10, 20, 30)),
  response_rates_list = list(c(0.1, 0.2, 0.3)),
  n_trials            = 10)

analyses_list <- performAnalyses(
  scenario_list       = scenarios_list,
  target_rates        = rep(0.5, 3),
  n_mcmc_iterations   = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 14-Apr-2026 Performing Analyses
#>          Analyzing 1 scenario (10 unique trial realizations)
#>    11:49 - with Berry ...........
#>              finished after 0 secs.
#>    11:49 - with Exnex ...........
#>              finished after 0.1 secs.
#>    11:49 - with Exnex_adj .......
#>              finished after 0.1 secs.
#>    11:49 - with Exnex_adj_mix ...
#>              finished after 0.1 secs.
#>    11:49 - with Exnex_mix .......
#>              finished after 0.1 secs.
#>    11:49 - with Pooled ..........
#>              finished after 0 secs.
#>    11:49 - with Stratified ......
#>              finished after 0 secs.
#>    11:49 - with Stratified_mix ...
#>              finished after 0 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

scaleRoundList(
  list         = getEstimates(analyses_list),
  scale_param  = 100,
  round_digits = 2)
#> $berry
#> $berry$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 17.52 7.97  4.37 16.74 34.04  6.74 0.88
#> p_2 18.45 6.21  8.04 17.86 31.81 -2.14 0.79
#> p_3 27.92 7.18 15.38 27.58 43.23 -2.42 0.89
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 15.78 9.08  3.69 13.99 36.73  3.99 0.63
#> p_2 18.23 7.52  6.53 17.28 33.81 -2.72 0.88
#> p_3 29.89 7.70 16.19 29.42 45.65 -0.58 1.04
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 17.06 9.56  3.50 15.67 39.36  5.67 0.80
#> p_2 18.62 7.13  7.36 18.00 33.66 -2.00 0.87
#> p_3 30.25 7.78 16.93 29.96 45.34 -0.04 1.08
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 16.17 9.53  2.81 14.46 38.08  4.46 0.62
#> p_2 18.39 7.50  6.38 17.72 34.03 -2.28 0.90
#> p_3 30.04 7.88 17.04 29.41 45.78 -0.59 1.02
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 15.81 9.01  3.18 14.32 35.80  4.32 0.57
#> p_2 18.40 7.20  6.51 17.67 33.39 -2.33 0.97
#> p_3 30.05 7.47 17.01 29.73 45.16 -0.27 0.94
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 22.95 5.28 13.51 22.65 34.07 12.65 1.94
#> p_2 22.95 5.28 13.51 22.65 34.07  2.65 0.41
#> p_3 22.95 5.28 13.51 22.65 34.07 -7.35 0.88
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 14.55 9.70  2.03 12.47 38.31  2.47 0.75
#> p_2 18.10 7.68  6.16 17.11 35.52 -2.89 1.20
#> p_3 30.97 7.98 16.64 30.55 47.62  0.55 1.05
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 15.08 9.86  2.46 12.77 36.86  2.77 0.72
#> p_2 18.38 7.72  7.03 17.53 35.81 -2.47 1.23
#> p_3 31.12 7.97 17.18 31.24 46.89  1.24 1.05
#> 
#>