scaleRoundList

This function applies scaling and rounding to each item of a list of numerics

scaleRoundList(list, scale_param = 1, round_digits = NULL)

Arguments

list

The list to which the scaling and rounding should be applied to.

scale_param

A numeric for the scaling of each item of the list, Default: `1`

round_digits

An integer for the number of digits. If `NULL`, no rounding will be applied, Default: `NULL`

Value

A list of scaled and rounded numerics

Author

Stephan Wojciekowski

Examples

some_list <- as.list(runif(5))
scaleRoundList(some_list, scale_param = 100, round_digits = 2)
#> [[1]]
#> [1] 84.31
#> 
#> [[2]]
#> [1] 6.49
#> 
#> [[3]]
#> [1] 76.21
#> 
#> [[4]]
#> [1] 7.84
#> 
#> [[5]]
#> [1] 13.35
#> 

scenarios_list <- simulateScenarios(
  n_subjects_list     = list(c(10, 20, 30)),
  response_rates_list = list(c(0.1, 0.2, 0.3)),
  n_trials            = 10)

analyses_list <- performAnalyses(
  scenario_list       = scenarios_list,
  target_rates        = rep(0.5, 3),
  n_mcmc_iterations   = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 24-Apr-2026 Performing Analyses
#>          Analyzing 1 scenario (10 unique trial realizations)
#>    12:43 - with Berry ...........
#>              finished after 0.1 secs.
#>    12:43 - with Exnex ...........
#>              finished after 0.1 secs.
#>    12:43 - with Exnex_adj .......
#>              finished after 0.1 secs.
#>    12:43 - with Exnex_adj_mix ...
#>              finished after 0.1 secs.
#>    12:43 - with Exnex_mix .......
#>              finished after 0.1 secs.
#>    12:43 - with Pooled ..........
#>              finished after 0.1 secs.
#>    12:43 - with Stratified ......
#>              finished after 0.1 secs.
#>    12:43 - with Stratified_mix ...
#>              finished after 0.1 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

scaleRoundList(
  list         = getEstimates(analyses_list),
  scale_param  = 100,
  round_digits = 2)
#> $berry
#> $berry$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 20.72 9.23  5.48 20.97 37.38 10.97 1.77
#> p_2 26.12 8.37 12.21 25.67 43.29  5.67 0.82
#> p_3 30.37 7.39 16.99 30.49 45.53  0.49 0.38
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>      Mean   SD  2.5%   50% 97.5% Bias  MSE
#> p_1 17.38 9.74  3.29 16.18 39.72 6.18 0.82
#> p_2 26.07 8.90 11.21 25.36 44.48 5.36 1.04
#> p_3 32.30 7.80 18.64 31.80 48.24 1.80 0.50
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>      Mean   SD  2.5%   50% 97.5% Bias  MSE
#> p_1 16.79 9.57  3.33 14.90 37.74 4.90 0.58
#> p_2 26.13 8.98 11.37 25.74 45.54 5.74 1.19
#> p_3 32.23 8.14 17.94 31.87 49.19 1.87 0.51
#> 
#> 
#> $exnex_adj_mix
#> $exnex_adj_mix$scenario_1
#>      Mean   SD  2.5%   50% 97.5% Bias  MSE
#> p_1 17.41 9.78  3.03 16.09 38.49 6.09 0.67
#> p_2 25.42 8.64 10.68 24.67 43.30 4.67 1.05
#> p_3 32.43 8.05 18.11 31.98 48.67 1.98 0.50
#> 
#> 
#> $exnex_mix
#> $exnex_mix$scenario_1
#>      Mean   SD  2.5%   50% 97.5% Bias  MSE
#> p_1 17.63 9.74  3.40 16.07 39.35 6.07 0.77
#> p_2 26.07 8.79 11.07 25.54 43.86 5.54 1.15
#> p_3 31.95 7.90 18.47 31.50 48.24 1.50 0.46
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 26.72 5.58 16.56 26.47 38.32 16.47 2.95
#> p_2 26.72 5.58 16.56 26.47 38.32  6.47 0.66
#> p_3 26.72 5.58 16.56 26.47 38.32 -3.53 0.37
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>      Mean   SD  2.5%   50% 97.5% Bias  MSE
#> p_1 14.55 9.79  1.87 12.47 38.49 2.47 0.55
#> p_2 25.71 9.03 10.49 24.94 45.28 4.94 1.29
#> p_3 33.23 8.23 18.26 32.86 50.23 2.86 0.61
#> 
#> 
#> $stratified_mix
#> $stratified_mix$scenario_1
#>      Mean   SD  2.5%   50% 97.5% Bias  MSE
#> p_1 15.09 9.94  2.26 13.47 37.25 3.47 0.66
#> p_2 25.97 9.03 10.61 25.36 44.68 5.36 1.41
#> p_3 33.36 8.21 19.29 32.64 49.09 2.64 0.64
#> 
#>