scaleRoundList

This function applies scaling and rounding to each item of a list of numerics

scaleRoundList(list, scale_param = 1, round_digits = NULL)

Arguments

list

The list to which the scaling and rounding should be applied to.

scale_param

A numeric for the scaling of each item of the list, Default: `1`

round_digits

An integer for the number of digits. If `NULL`, no rounding will be applied, Default: `NULL`

Value

A list of scaled and rounded numerics

Author

Stephan Wojciekowski

Examples

some_list <- as.list(runif(5))
scaleRoundList(some_list, scale_param = 100, round_digits = 2)
#> [[1]]
#> [1] 89.27
#> 
#> [[2]]
#> [1] 84.39
#> 
#> [[3]]
#> [1] 72.98
#> 
#> [[4]]
#> [1] 3.84
#> 
#> [[5]]
#> [1] 33.38
#> 

scenarios_list <- simulateScenarios(
  n_subjects_list     = list(c(10, 20, 30)),
  response_rates_list = list(c(0.1, 0.2, 0.3)),
  n_trials            = 10)

analyses_list <- performAnalyses(
  scenario_list       = scenarios_list,
  target_rates        = rep(0.5, 3),
  n_mcmc_iterations   = 100)
#> 
#> Caution: No parallel backend detected for the 'foreach' framework. For execution in parallel, register a parallel backend, e.g. with:
#>    doFuture::registerDoFuture()
#>    future::plan(future::multisession)
#> 20-Feb-2026 Performing Analyses
#>          Analyzing 1 scenario (10 unique trial realizations)
#>    17:00 - with Berry ...........
#>              finished after 0.1 secs.
#>    17:00 - with Exnex ...........
#>              finished after 0.1 secs.
#>    17:00 - with Exnex_adj .......
#>              finished after 0.1 secs.
#>    17:00 - with Pooled ..........
#>              finished after 0 secs.
#>    17:00 - with Stratified ......
#>              finished after 0 secs.
#>          Processing scenarios ...
#>              finished after 0 secs.

scaleRoundList(
  list         = getEstimates(analyses_list),
  scale_param  = 100,
  round_digits = 2)
#> $berry
#> $berry$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 16.84 7.90  4.54 16.06 34.65  6.06 0.74
#> p_2 17.93 6.56  7.24 17.04 31.85 -2.96 0.59
#> p_3 28.36 7.26 16.61 27.67 44.37 -2.33 1.08
#> 
#> 
#> $exnex
#> $exnex$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 16.53 8.67  4.06 15.15 35.64  5.15 0.74
#> p_2 18.74 7.85  6.82 17.76 36.62 -2.24 0.66
#> p_3 30.52 7.90 17.41 30.06 46.71  0.06 1.12
#> 
#> 
#> $exnex_adj
#> $exnex_adj$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 16.65 9.69  3.57 14.43 38.68  4.43 0.63
#> p_2 18.16 7.31  6.76 17.38 34.37 -2.62 0.73
#> p_3 31.13 7.88 17.31 30.71 47.04  0.71 1.08
#> 
#> 
#> $pooled
#> $pooled$scenario_1
#>      Mean   SD 2.5%   50% 97.5%  Bias  MSE
#> p_1 23.11 5.31 13.6 22.82 34.29 12.82 1.87
#> p_2 23.11 5.31 13.6 22.82 34.29  2.82 0.31
#> p_3 23.11 5.31 13.6 22.82 34.29 -7.18 0.74
#> 
#> 
#> $stratified
#> $stratified$scenario_1
#>      Mean   SD  2.5%   50% 97.5%  Bias  MSE
#> p_1 14.55 9.70  2.03 12.47 38.31  2.47 0.75
#> p_2 17.62 7.71  5.68 16.61 35.16 -3.39 0.95
#> p_3 31.61 8.02 17.17 31.21 48.29  1.21 1.13
#> 
#>